+++ CHANGES.md
@@ -9,6 +9,7 @@
 * Add `roundRobin` as a load balancer option.
 * Add the ability to route basted on the `dest` request header when using the
   TTwitter Thrift protocol
+* Add `preferServiceAddress` option to `io.l5d.consul` namer
 * Rename `baseDtab` router property to `dtab`
 * Add `io.l5d.adminMetricsExport` telemeter as an eventual replacement for the
   `io.l5d.commonMetrics` telemeter.